Backend user deletion and testing

+++ survey-backend/controllers/users.js
@@ -1,5 +1,7 @@
 const bcrypt = require('bcrypt')
+const jwt = require('jsonwebtoken')
 const router = require('express').Router()
+
 const User = require('../models/user')
 
 router.get('/', async (request, response) => {
@@ -32,6 +34,16 @@ router.post('/', async (request, response) => {
 })
 
 router.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  if (decodedToken.id !== request.params.id) {
+    return response.status(401).json({ error: 'only the account creator can delete the account' })
+  }
+
   await User.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
